// Wohnungsgeberbestätigung/src/Wohnungsgeberbestaetigung-Agent-Frontend/src/app/dto/meldebestaetigung-dto.ts
export class MeldebestaetigungDto {
  vorname: string;
  familienname: string;
  rufname: string;
  geborenAm: string;
  geburtsort: string;
  staatsangeh: string;
  famStand: string;
  religion: string;
  einzigeWohnung: string;
  einzugAm: string;
  gemeldetSeit: string;
  gemeindeKennziffer: string;
}